-lab_3 done

+++ src/lab_3/lab_3.hpp
@@ -74,7 +74,7 @@ struct Polygon {
 };
 
 struct ActiveEdge {
-    int currentX;
+    float currentX;
     float incrementX;
     int ymax;
